Disallow capturing a piece of the same color
I overlooked this case originally and it makes more sense
that this logic resides in the psuedo-legal move generators
as it's easy enough to check there.

To enable this to happen at the piece leve, all of the
psuedo-legal move functions now take a piece mapping instead
of a set, since they need to know about piece color in addition
to just location.

Took the time to cleanup some of the tests as well after
updating them to support this change.