Add en passant behavior
When calculating en passant as a psuedo-legal move, the `Move`
object is provided with a `captures` piece. This is in contrast
to other psuedo-legal generated moves, which let the legal move
generator determine if there was a piece that was captured during
the move.

Since the square where the capturing piece ends up is different
than the square of the captured piece for en passant, the `captures`
slot must be filled out during the move generation.

This required some changes to the make/unmake move functions, most
importantly around assumptions of which pieces are where when unmaking
a move.