Ensure transpositions are counted during perft
There isn't a need to check in the depth first search
if the current node was counted or not since it never
leaves the call stack; we only travel below it and then
back up. By adding the unncessary constraint that we'd only
count nodes of previously-unseen decendents, we avoided
counting all branches that contained a transposition.

Interestingly, for `test_perft_starting_position`, the node
counts for depth 3 and 4 were each ~23% lower than expected,
suggesting just how often transpositions can occur after only
a few moves!