Add apply and revert move functionality
Provides ability to (un)make an arbitrary `Move` to a
given `Pieces`. The `Pieces` internal representation
was changed to hold a pointer to each bitboard vector,
in turn requiring a bunch of syntax changes elsewhere
in the code.

The `{apply, revert}Move` logic is fairly dumb, in that
it doesn't check for any sort of chess legality before
updating `Pieces`. Rather, it is assumed that psuedo-
legal moves are provided (e.g. pieces are moving in
the proper fashion, etc).

Once castling rights, etc are added, it will likely
make sense for that logic to be checked for in these
functions.